FitSofia {SOfireA}	R Documentation

Fit a Sofia model to a data set

Description

The function fits a SOFIA model to a dataset.

Usage

FitSofia(x, y, unc = NULL, per.group = rep(FALSE, ncol(x)), nodes = 4, 
    sofiapar, restart = 0, cost = NULL, ...)

Arguments

x
data.frame with independent variables

y
dependent variable (observation)

unc
uncertainty of dependent variable

per.group
a boolean vector that indicates if a column in x acts per group (e.g. PFTs)

nodes
number of nodes for parallel compuation during genetic optimization

sofiapar
SofiaPar object with prior parameters

restart
restart previous Sofia optimization? 0 = start new, 1 = continue with previous, 2 = do post-processing

cost
cost function to be used

...
further arguments

Details

No details.

Author(s)

Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]

References

No reference.

See Also

Sofia, SofiaOpt

[Package SOfireA version 1.0 Index]