MakeFig {SOfireA}	R Documentation

Calculate figure positions for graphics that consist of multiple figure

Usage

MakeFig(nfig, border = c(0, 1, 0, 1), nrow = NULL, ncol = NULL)

Arguments

nfig
number of figures

border
relative graphic borders in which the figures should be placed

nrow
number of rows to arrange the figures

ncol
number of cols to arrange the figures

Value

A list with positions for each figure and number of rows and cloumns

Author(s)

Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]

[Package SOfireA version 1.0 Index]