SOfireA-package {SOfireA}	R Documentation

Satellite Observations for Fire Activity

Description

SOFIA (Satellite Observations for Fire Activity) is an empirical modelling concept to predict burned area based on satellite and climate data. The package implements the basic SOFIA model structure, and functions to optimize and plot SOFIA models.

Details

The DESCRIPTION file: This package was not yet installed at build time.

Index: This package was not yet installed at build time.

Author(s)

Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]

[Package SOfireA version 1.0 Index]