predict.Sofia {SOfireA}	R Documentation

Predict values based on a 'Sofia' object

Description

Make a predicition based on a Sofia object and newdata

Usage

## S3 method for class 'Sofia'
predict(object, newdata, return.all = FALSE, ...)

Arguments

object
an object of class 'Sofia', see Sofia

newdata
a data frame with columns names as in object$group.names for area fractions of groups and as in object$x.names for explantory variables

return.all
return all Sofia results? If FALSE, returns only total burned area

...
further arguments (not used)

Details

No details.

Value

A vector with predicted values.

Author(s)

Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]

References

No reference.

See Also

Sofia, SofiaOpt

[Package SOfireA version 1.0 Index]